labeled ctx done; no DOM label
Since we impose MAC on top of th SOP then liframe accesses DOM of
another we first check SOP then MAC. the SOP check is the same as the
DOM label. Hence, explanation is simpler this way. The fact that the
implementation changes the underlying origins when raising label and
we need a dom label to effectively track the origin is a separate
concern.